requirements.txt
Flask
Requests
python-dotenv